migrate methods to separate script
